+++ 1-exchanges/hareswap/src/lib.rs
@@ -3,10 +3,11 @@ mod interpreter;
 mod maker;
 mod requirement;
 mod transporter;
+mod model;
 
 /// The api used for off-ledger operations
 pub mod api {
-    pub use super::maker::*;
+    pub use super::model::*;
     pub use super::requirement::*;
     pub use super::transporter::authentication::{sign, verify, VerifyError};
     pub use super::transporter::decoder::*;
