+++ 1-exchanges/hareswap/src/transporter/voucher.rs
@@ -1,8 +1,13 @@
+use hex;
+
 use sbor::{describe::Type, Decode, DecodeError, Decoder, Describe, Encode, TypeId};
 use scrypto::prelude::*;
 
 use super::decoder::*;
+use super::authentication::*;
 
+/// The PassThruNFD implementation and related traits enables a NonFungibleData
+/// property to 
 #[derive(PartialEq, Eq, Debug, TypeId, Encode, Decode, Describe)]
 pub struct PassThruNFD {
     immutable_data: Vec<u8>,
@@ -41,12 +46,12 @@ impl NonFungibleData for PassThruNFD {
 
     /// Returns the schema of the immutable data.
     fn immutable_data_schema() -> Type {
-        panic!("unimplemented");
+        panic!("unimplemented"); // not needed to decode a PassThruNFD
     }
 
     /// Returns the schema of the mutable data.
     fn mutable_data_schema() -> Type {
-        panic!("unimplemented");
+        panic!("unimplemented"); // not needed to decode a PassThruNFD
     }
 }
 
@@ -54,9 +59,9 @@ impl NonFungibleData for PassThruNFD {
 
 #[derive(PartialEq, Eq, Debug, TypeId, Describe, Encode)]
 pub struct Voucher {
-    pub resource_def: ResourceDef,
-    pub key: Option<NonFungibleKey>,
-    pub nfd: PassThruNFD,
+    resource_def: ResourceDef,
+    key: Option<NonFungibleKey>,
+    nfd: PassThruNFD,
 }
 
 impl PrivateDecode for Voucher {
@@ -103,5 +108,40 @@ impl Voucher {
         };
         // finally mint
         resource_def.mint_non_fungible(&key, nfd, auth)
+    }
+
+
+    /// create a Voucher from any NonFungibleData with the the included metadata
+    pub fn from_nfd<T: NonFungibleData>(resource_def: ResourceDef, key: Option<NonFungibleKey>, nfd: T) -> Voucher {
+        Voucher {
+            resource_def,
+            key,
+            nfd: nfd.as_passthru(), // calling as_passthru is the "trick" and is an implementation detail for how Vouchers work which can be ignored for users.
+        }
+    }
+
+    /// create a SealedVoucher from this Voucher and an opaque signature
+    pub fn to_sealed(&self, signature: Vec<u8>) -> SealedVoucher {
+        SealedVoucher {
+            serialized: scrypto_encode(self),
+            signature
+        }
+    }
+}
+
+/// An opaque data structure which can (only) be converted back into a Voucher by validating the digital signature
+#[derive(TypeId, Describe, Encode, Decode)]
+pub struct SealedVoucher {
+    serialized: Vec<u8>,
+    signature: Vec<u8>,
+}
+
+impl SealedVoucher {
+    /// Converts a SealedVoucher back to a Voucher by verifying the signature against public_key
+    pub fn unseal(&self, public_key: &EcdsaPublicKey) -> Voucher {
+        debug!("SealedVoucher::unseal: serialized: {}", hex::encode(&self.serialized));
+        debug!("SealedVoucher::unseal:  signature: {}", hex::encode(&self.signature));
+        verify_or_panic(public_key, &self.serialized, &self.signature); // NOTE: panics on failure
+        private_decode_with_type(&self.serialized).unwrap()
     }
 }
